% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{backgroundIDs}
\alias{backgroundIDs}
\title{background IDs yeast}
\format{character vector}
\source{
Biomart, Ensembl 88
}
\usage{
backgroundIDs
}
\value{
character vector
}
\description{
example of background IDs
}
\author{
Thomas Schwarzl, 2017-04-11
}
