% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{foregroundIDs}
\alias{foregroundIDs}
\title{foreground IDs yeast}
\format{character vector}
\source{
Biomart, Ensembl 88
}
\usage{
foregroundIDs
}
\value{
character vector
}
\description{
example of foreground IDs
}
\author{
Thomas Schwarzl, 2017-04-11
}
