% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnrichment.R
\name{plotEnrichment}
\alias{plotEnrichment}
\title{Displaying enrichment}
\usage{
plotEnrichment(X, nameColumn, onlySignificant,
                         maxDisplay, verbose,
                         gg, las, oma, ...)
}
\arguments{
\item{X}{tibble. results table from the termEnrichment function}

\item{nameColumn}{character. Name of column for the plot to display}

\item{onlySignificant}{logical. display only signficant results if TRUE. significance is determined by column "significant"}

\item{maxDisplay}{numeric. maximum number of terms to display}

\item{verbose}{logical. hides messages if FALSE}

\item{gg}{logical. use ggplot2 if TRUE. dafault: TRUE}

\item{las}{numeric grapnical parameter las}

\item{oma}{vector graphical parameter oma}

\item{...}{graphical parameters}
}
\value{
ggplot object if gg = TRUE, else plots barchart
}
\description{
visualization Barplot of log2 odd ratios of significantly enriched RNA types
}
\examples{

data(yeastGO)
data(backgroundIDs)
data(foregroundIDs)
data(yeastGOdesc)

X <- termEnrichment(yeastGO, foregroundIDs, backgroundIDs, annotation = yeastGOdesc)
plotEnrichment(X)
}
\author{
Thomas Schwarzl <schwarzl@embl.de>
}
