% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeastGO}
\alias{yeastGO}
\title{yeast ids to GO annotation}
\format{\code{tibble} instance}
\source{
Biomart, Ensembl 88
}
\usage{
yeastGO
}
\value{
\code{tibble} instance
}
\description{
yeast ids to GO annotation
}
\author{
Thomas Schwarzl, 2017-04-11
}
