% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeastGOdesc}
\alias{yeastGOdesc}
\title{yeast GO term description}
\format{\code{tibble} instance}
\source{
Biomart, Ensembl 88
}
\usage{
yeastGOdesc
}
\value{
\code{tibble} instance
}
\description{
yeast GO term description
}
\author{
Thomas Schwarzl, 2017-04-11
}
