% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeastInterPro}
\alias{yeastInterPro}
\title{yeast ids to inter pro annotation}
\format{\code{tibble} instance}
\source{
Biomart, Ensembl 88
}
\usage{
yeastInterPro
}
\value{
\code{tibble} instance
}
\description{
yeast ids to inter pro annotation
}
\author{
Thomas Schwarzl, 2017-04-11
}
