% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeastInterProDesc}
\alias{yeastInterProDesc}
\title{yeast inter pro description}
\format{\code{tibble} instance}
\source{
Biomart, Ensembl 88
}
\usage{
yeastInterProDesc
}
\value{
\code{tibble} instance
}
\description{
yeast inter pro description
}
\author{
Thomas Schwarzl, 2017-04-11
}
