% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termEnrichment.R
\name{termEnrichment}
\alias{termEnrichment}
\title{termEnrichment}
\usage{
termEnrichment(termTable, foregroundIDs, backgroundIDs, annotation, terms, padj.cutoff, padj.method, alternative, parallel)
}
\arguments{
\item{termTable}{data.frame or tibble ID where first column is an ID column and the second column are associated terms to be tested against}

\item{foregroundIDs}{vector of IDs defining the foreground (must be same IDs as is first column of termTable)}

\item{backgroundIDs}{vector of IDs defining the background (must be same IDs as is first column of termTable)}

\item{annotation}{data.frame or tibble annotating terms (first column must be term identifier)}

\item{terms}{specifies what terms will be tested for. tests all existing terms if NULL}

\item{padj.cutoff}{cutoff for p.adjusted value}

\item{padj.method}{p adjustment method for fisher exact test, see ?fisher.test}

\item{alternative}{alternative hypothesis for fisher exact test, see ?fisher.test}

\item{parallel}{logical the testing be parallised with BiocParallel}
}
\value{
tibble annotated analysis results
}
\description{
This function calculates enrichment for a set of features against a set
of background features.
}
\examples{

data(yeastGO)
data(backgroundIDs)
data(foregroundIDs)
data(yeastGOdesc)

termEnrichment(yeastGO, foregroundIDs, backgroundIDs,
               annotation = yeastGOdesc)
}
\author{
Thomas Schwarzl <schwarzl@embl.de>
}
